fix: no more twig runtime in clientHelperBundle
And all attribute extension
