+++ EMS/client-helper-bundle/config/services.php
@@ -15,12 +15,10 @@ use EMS\ClientHelperBundle\Controller\SearchController;
 use EMS\ClientHelperBundle\EventListener\CacheListener;
 use EMS\ClientHelperBundle\EventListener\KernelListener;
 use EMS\ClientHelperBundle\EventListener\SecurityListener;
-use EMS\ClientHelperBundle\Helper\Asset\AssetHelperRuntime;
 use EMS\ClientHelperBundle\Helper\Asset\AssetVersionStrategy;
 use EMS\ClientHelperBundle\Helper\Cache\CacheHelper;
 use EMS\ClientHelperBundle\Helper\ContentType\ContentTypeHelper;
 use EMS\ClientHelperBundle\Helper\Elasticsearch\ClientRequestManager;
-use EMS\ClientHelperBundle\Helper\Elasticsearch\ClientRequestRuntime;
 use EMS\ClientHelperBundle\Helper\Environment\EnvironmentFactory;
 use EMS\ClientHelperBundle\Helper\Environment\EnvironmentHelper;
 use EMS\ClientHelperBundle\Helper\Form\Extension\EmschFormViewExtension;
@@ -30,8 +28,8 @@ use EMS\ClientHelperBundle\Helper\Request\ExceptionHelper;
 use EMS\ClientHelperBundle\Helper\Request\LocaleHelper;
 use EMS\ClientHelperBundle\Helper\Translation\Translator;
 use EMS\ClientHelperBundle\Helper\Webhook\WebhookHelper;
-use EMS\ClientHelperBundle\Helper\Webhook\WebhookRuntime;
 use EMS\ClientHelperBundle\Twig\AdminMenuExtension;
+use EMS\ClientHelperBundle\Twig\AssetExtension;
 use EMS\ClientHelperBundle\Twig\HelperExtension;
 use EMS\CommonBundle\Contracts\Bridge\Core\CoreBridgeInterface;
 use EMS\CommonBundle\Contracts\Elasticsearch\QueryLoggerInterface;
@@ -109,10 +107,6 @@ return static function (ContainerConfigurator $container) {
             service('ems.common.cache'),
         ]);
 
-    $services->set('emsch.runtime_webhook', WebhookRuntime::class)
-        ->args([service('emsch.helper_webhook')])
-        ->tag('twig.runtime');
-
     $services->set('emsch.form.extension.view', EmschFormViewExtension::class)
         ->tag('form.type_extension', ['priority' => 1]);
 
@@ -148,14 +142,24 @@ return static function (ContainerConfigurator $container) {
         ->tag('kernel.event_subscriber');
 
     $services->set('emsch.twig.extension.helper', HelperExtension::class)
-        ->tag('twig.extension');
+        ->args([
+            service('emsch.manager.client_request'),
+            service('request_stack'),
+            service('logger'),
+            service('ems_common.service.elastica'),
+            service('emsch.search.manager'),
+            service('emsch.helper_webhook'),
+        ])
+        ->tag('monolog.logger', ['channel' => 'emsch_request'])
+        ->tag('twig.attribute_extension')
+        ->tag('twig.runtime');
 
     $services->set('emsch.twig.admin_menu_extension', AdminMenuExtension::class)
         ->args([service('emsch.helper_environment')])
         ->tag('twig.attribute_extension')
         ->tag('twig.runtime');
 
-    $services->set('emsch.twig.runtime.asset', AssetHelperRuntime::class)
+    $services->set('emsch.twig.asset_extension', AssetExtension::class)
         ->args([
             service('ems_common.storage.manager'),
             service('ems_common.twig.runtime.asset'),
@@ -163,21 +167,11 @@ return static function (ContainerConfigurator $container) {
             '%kernel.project_dir%',
             '%emsch.asset_local_folder%',
         ])
+        ->tag('twig.attribute_extension')
         ->tag('twig.runtime');
 
-    $services->set('emsch.twig.runtime.asset_vserion_strategy', AssetVersionStrategy::class)
-        ->args([service('emsch.twig.runtime.asset')]);
-
-    $services->set('emsch.twig.runtime.client_request', ClientRequestRuntime::class)
-        ->args([
-            service('emsch.manager.client_request'),
-            service('request_stack'),
-            service('logger'),
-            service('ems_common.service.elastica'),
-            service('emsch.search.manager'),
-        ])
-        ->tag('monolog.logger', ['channel' => 'emsch_request'])
-        ->tag('twig.runtime');
+    $services->set('emsch.asset.version_strategy', AssetVersionStrategy::class)
+        ->args([service('emsch.twig.asset_extension')]);
 
     $services->set(CoreBridgeController::class)
         ->public()
