--- EMS/client-helper-bundle/src/Helper/Webhook/WebhookRuntime.php
@@ -1,17 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace EMS\ClientHelperBundle\Helper\Webhook;
-
-class WebhookRuntime
-{
-    public function __construct(private readonly WebhookHelper $webhookHelper)
-    {
-    }
-
-    public function getWebhook(): Webhook
-    {
-        return $this->webhookHelper->getWebhook();
-    }
-}
