+++ elasticms-web/config/packages/assets.yaml
@@ -4,4 +4,4 @@ framework:
             emsform:
                 json_manifest_path: '%kernel.project_dir%/public/bundles/emsform/manifest.json'
             emsch:
-                version_strategy: 'emsch.twig.runtime.asset_vserion_strategy'
+                version_strategy: 'emsch.asset.version_strategy'
